+++ packages/core/src/StateMachine.ts
@@ -47,12 +47,13 @@ import type {
   StateValue,
   TransitionDefinition,
   UnknownActionObject,
-  ResolvedStateMachineTypes
+  ResolvedStateMachineTypes,
+  StateSchema,
+  SnapshotStatus
 } from './types.ts';
 import { resolveReferencedActor, toStatePath } from './utils.ts';
 
-export const STATE_IDENTIFIER = '#';
-export const WILDCARD = '*';
+const STATE_IDENTIFIER = '#';
 
 export class StateMachine<
   TContext extends MachineContext,
@@ -67,7 +68,8 @@ export class StateMachine<
   TInput,
   TOutput,
   TEmitted extends EventObject = EventObject, // TODO: remove default
-  TMeta extends MetaObject = MetaObject
+  TMeta extends MetaObject = MetaObject,
+  TConfig extends StateSchema = StateSchema
 > implements
     ActorLogic<
       MachineSnapshot<
@@ -77,7 +79,8 @@ export class StateMachine<
         TStateValue,
         TTag,
         TOutput,
-        TMeta
+        TMeta,
+        TConfig
       >,
       TEvent,
       TInput,
@@ -214,7 +217,7 @@ export class StateMachine<
       value: StateValue;
       context?: TContext;
       historyValue?: HistoryValue<TContext, TEvent>;
-      status?: 'active' | 'done' | 'error' | 'stopped';
+      status?: SnapshotStatus;
       output?: TOutput;
       error?: unknown;
     } & (Equals<TContext, MachineContext> extends false
@@ -227,7 +230,8 @@ export class StateMachine<
     TStateValue,
     TTag,
     TOutput,
-    TMeta
+    TMeta,
+    TConfig
   > {
     const resolvedStateValue = resolveStateValue(this.root, config.value);
     const nodeSet = getAllStateNodes(
@@ -254,7 +258,8 @@ export class StateMachine<
       TStateValue,
       TTag,
       TOutput,
-      TMeta
+      TMeta,
+      TConfig
     >;
   }
 
@@ -273,7 +278,8 @@ export class StateMachine<
       TStateValue,
       TTag,
       TOutput,
-      TMeta
+      TMeta,
+      TConfig
     >,
     event: TEvent,
     actorScope: ActorScope<typeof snapshot, TEvent, AnyActorSystem, TEmitted>
@@ -284,7 +290,8 @@ export class StateMachine<
     TStateValue,
     TTag,
     TOutput,
-    TMeta
+    TMeta,
+    TConfig
   > {
     return macrostep(snapshot, event, actorScope, [])
       .snapshot as typeof snapshot;
@@ -305,7 +312,8 @@ export class StateMachine<
       TStateValue,
       TTag,
       TOutput,
-      TMeta
+      TMeta,
+      TConfig
     >,
     event: TEvent,
     actorScope: AnyActorScope
@@ -317,7 +325,8 @@ export class StateMachine<
       TStateValue,
       TTag,
       TOutput,
-      TMeta
+      TMeta,
+      TConfig
     >
   > {
     return macrostep(snapshot, event, actorScope, []).microstates;
@@ -331,7 +340,8 @@ export class StateMachine<
       TStateValue,
       TTag,
       TOutput,
-      TMeta
+      TMeta,
+      TConfig
     >,
     event: TEvent
   ): Array<TransitionDefinition<TContext, TEvent>> {
@@ -353,7 +363,8 @@ export class StateMachine<
     TStateValue,
     TTag,
     TOutput,
-    TMeta
+    TMeta,
+    TConfig
   > {
     const { context } = this.config;
 
@@ -397,7 +408,8 @@ export class StateMachine<
         TStateValue,
         TTag,
         TOutput,
-        TMeta
+        TMeta,
+        TConfig
       >,
       TEvent,
       AnyActorSystem,
@@ -411,7 +423,8 @@ export class StateMachine<
     TStateValue,
     TTag,
     TOutput,
-    TMeta
+    TMeta,
+    TConfig
   > {
     const initEvent = createInitEvent(input) as unknown as TEvent; // TODO: fix;
     const internalQueue: AnyEventObject[] = [];
@@ -456,7 +469,8 @@ export class StateMachine<
       TStateValue,
       TTag,
       TOutput,
-      TMeta
+      TMeta,
+      TConfig
     >
   ): void {
     Object.values(snapshot.children as Record<string, AnyActorRef>).forEach(
@@ -500,7 +514,8 @@ export class StateMachine<
       TStateValue,
       TTag,
       TOutput,
-      TMeta
+      TMeta,
+      TConfig
     >,
     options?: unknown
   ) {
@@ -517,7 +532,8 @@ export class StateMachine<
         TStateValue,
         TTag,
         TOutput,
-        TMeta
+        TMeta,
+        TConfig
       >,
       TEvent,
       AnyActorSystem,
@@ -530,7 +546,8 @@ export class StateMachine<
     TStateValue,
     TTag,
     TOutput,
-    TMeta
+    TMeta,
+    TConfig
   > {
     const children: Record<string, AnyActorRef> = {};
     const snapshotChildren: Record<
@@ -584,7 +601,8 @@ export class StateMachine<
       TStateValue,
       TTag,
       TOutput,
-      TMeta
+      TMeta,
+      TConfig
     >;
 
     let seen = new Set();