+++ packages/core/src/index.ts
@@ -1,49 +1,41 @@
-export * from './actions.ts';
-export * from './actors/index.ts';
 export { SimulatedClock } from './SimulatedClock.ts';
-export { type Spawner } from './spawn.ts';
 export { isMachineSnapshot, type MachineSnapshot } from './State.ts';
 export { StateMachine } from './StateMachine.ts';
-export { getStateNodes } from './stateUtils.ts';
-export * from './types.ts';
-export { waitFor } from './waitFor.ts';
-import { createMachine } from './createMachine.ts';
-export { getInitialSnapshot, getNextSnapshot } from './getNextSnapshot.ts';
-import {
+export { StateNode } from './StateNode.ts';
+export * from './actions.ts';
+export * from './actors/index.ts';
+export { assertEvent } from './assert.ts';
+export {
   Actor,
   createActor,
   interpret,
   type Interpreter
 } from './createActor.ts';
-import { StateNode } from './StateNode.ts';
 export { executeAction } from './stateUtils.ts';
 // TODO: decide from where those should be exported
+export { createMachine } from './createMachine.ts';
+export { getInitialSnapshot, getNextSnapshot } from './getNextSnapshot.ts';
 export { and, not, or, stateIn } from './guards.ts';
-export { setup } from './setup.ts';
-export type { ActorSystem } from './system.ts';
 export type {
   InspectedActorEvent,
   InspectedEventEvent,
   InspectedSnapshotEvent,
   InspectionEvent
 } from './inspection.ts';
+export { setup } from './setup.ts';
+export { type Spawner } from './spawn.ts';
+export { getStateNodes } from './stateUtils.ts';
+export type { ActorSystem } from './system.ts';
 export { toPromise } from './toPromise.ts';
+export * from './types.ts';
 export {
   getAllOwnEventDescriptors as __unsafe_getAllOwnEventDescriptors,
   matchesState,
   pathToStateValue,
   toObserver
 } from './utils.ts';
-export {
-  Actor,
-  createActor,
-  createMachine,
-  interpret,
-  StateNode,
-  type Interpreter
-};
-export { assertEvent } from './assert.ts';
 export { transition } from './transition.ts';
+export { waitFor } from './waitFor.ts';
 
 declare global {
   interface SymbolConstructor {