+++ packages/core/src/stateUtils.ts
@@ -60,7 +60,7 @@ type AnyStateNodeIterable = StateNodeIterable<any, any>;
 
 type AdjList = Map<AnyStateNode, Array<AnyStateNode>>;
 
-export const isAtomicStateNode = (stateNode: StateNode<any, any>) =>
+const isAtomicStateNode = (stateNode: StateNode<any, any>) =>
   stateNode.type === 'atomic' || stateNode.type === 'final';
 
 function getChildren<TContext extends MachineContext, TE extends EventObject>(
@@ -160,10 +160,9 @@ function getValueFromAdj(baseNode: AnyStateNode, adjList: AdjList): StateValue {
   return stateValue;
 }
 
-export function getAdjList<
-  TContext extends MachineContext,
-  TE extends EventObject
->(stateNodes: StateNodeIterable<TContext, TE>): AdjList {
+function getAdjList<TContext extends MachineContext, TE extends EventObject>(
+  stateNodes: StateNodeIterable<TContext, TE>
+): AdjList {
   const adjList: AdjList = new Map();
 
   for (const s of stateNodes) {
@@ -486,7 +485,7 @@ export function formatInitialTransition<
   return transition;
 }
 
-export function resolveTarget(
+function resolveTarget(
   stateNode: AnyStateNode,
   targets: ReadonlyArray<string | AnyStateNode> | undefined
 ): ReadonlyArray<AnyStateNode> | undefined {
@@ -552,9 +551,7 @@ function isHistoryNode(
   return stateNode.type === 'history';
 }
 
-export function getInitialStateNodesWithTheirAncestors(
-  stateNode: AnyStateNode
-) {
+function getInitialStateNodesWithTheirAncestors(stateNode: AnyStateNode) {
   const states = getInitialStateNodes(stateNode);
   for (const initialState of states) {
     for (const ancestor of getProperAncestors(initialState, stateNode)) {
@@ -586,10 +583,7 @@ export function getInitialStateNodes(stateNode: AnyStateNode) {
   return set;
 }
 /** Returns the child state node from its relative `stateKey`, or throws. */
-export function getStateNode(
-  stateNode: AnyStateNode,
-  stateKey: string
-): AnyStateNode {
+function getStateNode(stateNode: AnyStateNode, stateKey: string): AnyStateNode {
   if (isStateId(stateKey)) {
     return stateNode.machine.getStateNodeById(stateKey);
   }
@@ -677,7 +671,7 @@ export function getStateNodes<
   );
 }
 
-export function transitionAtomicNode<
+function transitionAtomicNode<
   TContext extends MachineContext,
   TEvent extends EventObject
 >(
@@ -704,7 +698,7 @@ export function transitionAtomicNode<
   return next;
 }
 
-export function transitionCompoundNode<
+function transitionCompoundNode<
   TContext extends MachineContext,
   TEvent extends EventObject
 >(
@@ -738,7 +732,7 @@ export function transitionCompoundNode<
   return next;
 }
 
-export function transitionParallelNode<
+function transitionParallelNode<
   TContext extends MachineContext,
   TEvent extends EventObject
 >(
@@ -852,7 +846,7 @@ function hasIntersection<T>(s1: Iterable<T>, s2: Iterable<T>): boolean {
   return false;
 }
 
-export function removeConflictingTransitions(
+function removeConflictingTransitions(
   enabledTransitions: Array<AnyTransitionDefinition>,
   stateNodeSet: Set<AnyStateNode>,
   historyValue: AnyHistoryValue
@@ -1867,7 +1861,7 @@ export function resolveStateValue(
   return getStateValue(rootNode, [...allStateNodes]);
 }
 
-export function stateValuesEqual(
+function stateValuesEqual(
   a: StateValue | undefined,
   b: StateValue | undefined
 ): boolean {